Simplified {Ipv4Network, Ipv6Network}::new method and make clippy happy

+++ src/ipv4_network.rs
@@ -39,7 +39,7 @@ impl Ipv4Network {
             return Err(IpNetworkError::NetmaskError(netmask));
         }
 
-        if u32::from(network_address).trailing_zeros() < (Self::LENGTH as u32 - netmask as u32) {
+        if u32::from(network_address).trailing_zeros() < u32::from(Self::LENGTH - netmask) {
             return Err(IpNetworkError::HostBitsSet);
         }
